+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - Increase phpstan analysis level to 4 and fix all issues !1672 @ChrisOelmueller
 - Introduce some more specific store gateway functionality !1686 @ChrisOelmueller
 - Replace Vue filters with prototypes !1689 @ChrisOelmueller
+- Remove unused Xhr methods for region applications !1707 @alex.simm
 
 ## Dev/Test/CI stuff
 - Fix banana unit test !1649 !1656 #964 @alex.simm